adds support for 2nd gen authentication token
The platform API made an update to allow direct usage of the app secret
as a bearer token, provided the new form of token with embedded security
enhancements was used. To that end, the client is updated with a flag
for forcing the legacy token exchange or uses the new token model.

The client library will return an error if a user attempts to use the
legacy style app id and secret with the new token model. New tokens have
a prefix of "cbkey_" to show they are the new format.