removes requirement for deprecated application_id
The application_id is no longer offered through the portal, leading to
confusion about provider configuration. This change removes the
requirement for the application_id.

If anyone uses the forced token swap and leaves off the application_id,
there isn't a clear error message for that bad configuration.
It is believed that no one is doing the old token swap mechanic anymore.